+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspace-title/index.tsx
